Move InteractsWithRedis to Illuminate\Foundation\Testing (#21967)
This would allow the class to be reused not only for Laravel testing but
project that needs to depends on redis directly.
