+++ tests/Redis/ConcurrentLimiterTest.php
@@ -5,6 +5,7 @@ namespace Illuminate\Tests\Redis;
 use PHPUnit\Framework\TestCase;
 use Illuminate\Redis\Limiters\ConcurrencyLimiter;
 use Illuminate\Contracts\Redis\LimiterTimeoutException;
+use Illuminate\Foundation\Testing\Concerns\InteractsWithRedis;
 
 /**
  * @group redislimiters
